+++ decide/census/urls.py
@@ -5,4 +5,5 @@ from . import views
 urlpatterns = [
     path('', views.CensusCreate.as_view(), name='census_create'),
     path('<int:voting_id>/', views.CensusDetail.as_view(), name='census_detail'),
+    path('peticion/', views.peticionCenso, name="peticion"),
 ]
