+++ decide/decide/settings.py
@@ -44,7 +44,7 @@ INSTALLED_APPS = [
     'rest_framework.authtoken',
     'rest_framework_swagger',
     'gateway',
-    'peticion'
+    
 ]
 
 REST_FRAMEWORK = {
