+++ decide/decide/urls.py
@@ -25,7 +25,7 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('doc/', schema_view),
     path('gateway/', include('gateway.urls')),
-    path('peticion/', include('peticion.urls')),
+    path('census/', include('census.urls')),
 ]
 
 for module in settings.MODULES:
