--- decide/peticion/urls.py
@@ -1,10 +1,0 @@
-from django.urls import path
-from . import views
-from django.conf import settings
-
-urlpatterns = [
-
-    path('', views.peticionCenso, name="peticion"),
-
-
-]
